Support additional jQuery instances loaded after widget
When another jQuery instance is loaded after the Redactor widget,
`window.jQuery` is clobbered with a jQuery *without* the Redactor
plugin. This happens commonly in complex admin forms, where another
widget includes its own version of jQuery. Since our event handlers
refer to `window.jQuery` at the time the event handler is run, and not
the time the event handler was bound, our event handlers use a jQuery
without the Redactor plugin.

Wrap the event handler initialization in a closure that retains the
value of `window.jQuery` at the time the Redactor widget is loaded.